Add the ability to have a static field in logs
I need to have some data on all log messages -
such as app name and environment. This changes
the formatter a little bit so that the user can
pass in a map (or, I guess, a prop list) the
contents of which will be included in all
outputted log lines.